requirements.txt
fastapi
uvicorn[standard]
python-socketio
azure-messaging-webpubsub-socketio
python-dotenv
gunicorn

